поддержка зп оружейки для чн

+++ src/xrGameCS/WeaponMagazinedWGrenade.h
@@ -78,7 +78,7 @@ public:
 	CWeaponAmmo*			m_pAmmo2;
 	shared_str				m_ammoSect2;
 	xr_vector<shared_str>	m_ammoTypes2;
-	u32						m_ammoType2;
+	u8						m_ammoType2;
 	shared_str				m_ammoName2;
 	int						iMagazineSize2;
 	xr_vector<CCartridge>	m_magazine2;
